point

+++ src/main/java/com/example/finalproject/delivery/repository/DeliveryRepository.java
@@ -3,6 +3,6 @@ package com.example.finalproject.delivery.repository;
 import com.example.finalproject.delivery.domain.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeliveryRepository extends JpaRepository<Delivery,Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery,Long> {
 
 }
